15 clean up sortableRecord
